+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -3,15 +3,18 @@ import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
 import { socketService } from "~/libs/modules/socket/socket.js";
 import { token } from "~/libs/modules/token/token.js";
+import { activityController } from "~/modules/activities/activities.js";
 import { authController } from "~/modules/auth/auth.js";
 import { chatMessageController } from "~/modules/chat-messages/chat-messages.js";
 import { chatController } from "~/modules/chats/chats.js";
+import { commentController } from "~/modules/comments/comments.js";
 import { courseSectionController } from "~/modules/course-sections/course-sections.js";
 import { courseController } from "~/modules/courses/courses.js";
 import { fileController } from "~/modules/files/files.js";
 import { friendController } from "~/modules/friends/friends.js";
 import { sectionStatusController } from "~/modules/section-statuses/section-statuses.js";
 import { userCourseController } from "~/modules/user-courses/user-courses.js";
+import { userNotificationController } from "~/modules/user-notifications/user-notifications.js";
 import { userController, userService } from "~/modules/users/users.js";
 import { vendorController } from "~/modules/vendors/vendors.js";
 
@@ -21,6 +24,7 @@ import { BaseServerApplicationApi } from "./base-server-application-api.js";
 const apiV1 = new BaseServerApplicationApi(
 	"v1",
 	config,
+	...activityController.routes,
 	...authController.routes,
 	...courseController.routes,
 	...chatController.routes,
@@ -30,8 +34,10 @@ const apiV1 = new BaseServerApplicationApi(
 	...fileController.routes,
 	...friendController.routes,
 	...userController.routes,
+	...userNotificationController.routes,
 	...courseSectionController.routes,
 	...sectionStatusController.routes,
+	...commentController.routes,
 );
 const serverApplication = new BaseServerApplication({
 	apis: [apiV1],