+++ apps/backend/src/modules/comments/libs/validation-schemas/validation-schemas.ts
@@ -1,0 +1,6 @@
+export {
+	commentCreateBodyValidationSchema,
+	commentGetAllQueryValidationSchema,
+	commentIdParameterValidationSchema,
+	commentTextValidationSchema,
+} from "@trackmates/shared";