+++ apps/backend/src/modules/friends/friend.service.ts
@@ -1,6 +1,10 @@
 import { convertPageToZeroIndexed } from "~/libs/helpers/helpers.js";
 import { type PaginationResponseDto } from "~/libs/types/types.js";
 import { type FriendRepository } from "~/modules/friends/friend.repository.js";
+import {
+	type NotificationService,
+	NotificationType,
+} from "~/modules/notifications/notifications.js";
 import { type UserAuthResponseDto } from "~/modules/users/users.js";
 
 import {
@@ -11,9 +15,14 @@ import {
 
 class FriendService {
 	private friendRepository: FriendRepository;
+	private notificationService: NotificationService;
 
-	public constructor(friendRepository: FriendRepository) {
+	public constructor(
+		friendRepository: FriendRepository,
+		notificationService: NotificationService,
+	) {
 		this.friendRepository = friendRepository;
+		this.notificationService = notificationService;
 	}
 
 	public async createSubscription(
@@ -43,6 +52,12 @@ class FriendService {
 		const followingUser = await this.friendRepository.create({
 			followerUserId,
 			followingUserId,
+		});
+
+		await this.notificationService.create({
+			receiverUserId: followingUserId,
+			type: NotificationType.NEW_FOLLOWER,
+			userId: followerUserId,
 		});
 
 		return followingUser.toObject();