+++ apps/backend/src/modules/users/user.repository.ts
@@ -1,8 +1,8 @@
+import { RelationName } from "~/libs/enums/enums.js";
 import { type Repository } from "~/libs/types/types.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
 
-import { RelationName } from "./libs/enums/enums.js";
 import { type UserProfileRequestDto } from "./libs/types/types.js";
 import { type UserDetailsModel } from "./user-details.model.js";