+++ apps/frontend/src/libs/components/app/app.tsx
@@ -9,6 +9,7 @@ import {
 import { actions as appActions } from "~/libs/slices/app/app.js";
 import { actions as authActions } from "~/modules/auth/auth.js";
 import { actions as chatActions } from "~/modules/chats/chats.js";
+import { actions as userNotificationsActions } from "~/modules/user-notifications/user-notifications.js";
 
 import { Loader } from "../loader/loader.js";
 
@@ -26,6 +27,12 @@ const App: React.FC = () => {
 	useEffect(() => {
 		void dispatch(authActions.getAuthenticatedUser());
 	}, [dispatch]);
+
+	useEffect(() => {
+		if (user) {
+			void dispatch(userNotificationsActions.checkHasUserUnreadNotifications());
+		}
+	}, [dispatch, user]);
 
 	useEffect(() => {
 		if (redirectTo) {