+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -10,12 +10,17 @@ import { type Config } from "~/libs/modules/config/config.js";
 import { notification } from "~/libs/modules/notification/notification.js";
 import { socket } from "~/libs/modules/socket/socket.js";
 import { reducer as appReducer } from "~/libs/slices/app/app.js";
+import {
+	activitiesApi,
+	reducer as activitiesReducer,
+} from "~/modules/activities/activities.js";
 import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
 import {
 	chatMessagesApi,
 	reducer as chatMessagesReducer,
 } from "~/modules/chat-messages/chat-messages.js";
 import { chatsApi, reducer as chatsReducer } from "~/modules/chats/chats.js";
+import { commentApi } from "~/modules/comments/comments.js";
 import {
 	courseSectionsApi,
 	reducer as courseSectionsReducer,
@@ -37,6 +42,10 @@ import {
 	userCourseApi,
 	reducer as userCoursesReducer,
 } from "~/modules/user-courses/user-courses.js";
+import {
+	userNotificationsApi,
+	reducer as userNotificationsReducer,
+} from "~/modules/user-notifications/user-notifications.js";
 import { userApi, reducer as usersReducer } from "~/modules/users/users.js";
 import {
 	vendorApi,
@@ -47,6 +56,7 @@ import { storage } from "../storage/storage.js";
 import { chatSocket, handleError } from "./middlewares/middlewares.js";
 
 type RootReducer = {
+	activities: ReturnType<typeof activitiesReducer>;
 	app: ReturnType<typeof appReducer>;
 	auth: ReturnType<typeof authReducer>;
 	chatMessages: ReturnType<typeof chatMessagesReducer>;
@@ -56,14 +66,17 @@ type RootReducer = {
 	friends: ReturnType<typeof friendsReducer>;
 	sectionStatuses: ReturnType<typeof sectionStatusesReducer>;
 	userCourses: ReturnType<typeof userCoursesReducer>;
+	userNotifications: ReturnType<typeof userNotificationsReducer>;
 	users: ReturnType<typeof usersReducer>;
 	vendors: ReturnType<typeof vendorsReducer>;
 };
 
 type ExtraArguments = {
+	activitiesApi: typeof activitiesApi;
 	authApi: typeof authApi;
 	chatMessagesApi: typeof chatMessagesApi;
 	chatsApi: typeof chatsApi;
+	commentApi: typeof commentApi;
 	courseApi: typeof courseApi;
 	courseSectionsApi: typeof courseSectionsApi;
 	filesApi: typeof filesApi;
@@ -74,6 +87,7 @@ type ExtraArguments = {
 	storage: typeof storage;
 	userApi: typeof userApi;
 	userCourseApi: typeof userCourseApi;
+	userNotificationsApi: typeof userNotificationsApi;
 	vendorApi: typeof vendorApi;
 };
 
@@ -97,6 +111,7 @@ class Store {
 				}).prepend([handleError, chatSocket({ extra: this.extraArguments })]);
 			},
 			reducer: {
+				activities: activitiesReducer,
 				app: appReducer,
 				auth: authReducer,
 				chatMessages: chatMessagesReducer,
@@ -106,6 +121,7 @@ class Store {
 				friends: friendsReducer,
 				sectionStatuses: sectionStatusesReducer,
 				userCourses: userCoursesReducer,
+				userNotifications: userNotificationsReducer,
 				users: usersReducer,
 				vendors: vendorsReducer,
 			},
@@ -114,9 +130,11 @@ class Store {
 
 	public get extraArguments(): ExtraArguments {
 		return {
+			activitiesApi,
 			authApi,
 			chatMessagesApi,
 			chatsApi,
+			commentApi,
 			courseApi,
 			courseSectionsApi,
 			filesApi,
@@ -127,6 +145,7 @@ class Store {
 			storage,
 			userApi,
 			userCourseApi,
+			userNotificationsApi,
 			vendorApi,
 		};
 	}