+++ apps/frontend/src/pages/chats/libs/components/chat-sidebar/chat-sidebar.tsx
@@ -4,6 +4,7 @@ import { AppRoute } from "~/libs/enums/app-route.enum.js";
 import { type ChatGetAllItemResponseDto } from "~/modules/chats/chats.js";
 
 import { ChatLink } from "../chat-link/chat-link.js";
+import { ChatSearchBar } from "../chat-search-bar/chat-search-bar.js";
 import styles from "./styles.module.css";
 
 type Properties = {
@@ -14,26 +15,33 @@ const ChatSidebar: React.FC<Properties> = ({ chats }: Properties) => {
 	const hasChats = chats.length > EMPTY_ARRAY_LENGTH;
 
 	return (
-		<ul className={styles["chats"]}>
-			{hasChats &&
-				chats.map((item) => {
-					return (
-						<li key={item.id}>
-							<ChatLink chat={item} />
+		<>
+			<div className={styles["container"]}>
+				<div className={styles["search-container"]}>
+					<ChatSearchBar />
+				</div>
+				<ul className={styles["chats"]}>
+					{hasChats &&
+						chats.map((item) => {
+							return (
+								<li key={item.id}>
+									<ChatLink chat={item} />
+								</li>
+							);
+						})}
+					{!hasChats && (
+						<li>
+							<p className={styles["empty-chat-list"]}>
+								There are no chats yet.{" "}
+								<Link className={styles["link"]} to={AppRoute.FRIENDS}>
+									Create new one.
+								</Link>
+							</p>
 						</li>
-					);
-				})}
-			{!hasChats && (
-				<li>
-					<p className={styles["empty-chat-list"]}>
-						There are no chats yet.{" "}
-						<Link className={styles["link"]} to={AppRoute.FRIENDS}>
-							Create new one.
-						</Link>
-					</p>
-				</li>
-			)}
-		</ul>
+					)}
+				</ul>
+			</div>
+		</>
 	);
 };