+++ apps/frontend/src/pages/overview/libs/components/welcome-section-header/welcome-section-header.tsx
@@ -34,6 +34,7 @@ const WelcomeHeader: React.FC<Properties> = ({
 			<div className={styles["welcome-section-header-actions"]}>
 				<Button
 					className={styles["add-course-button"]}
+					iconClassName={styles["add-course-button-icon"]}
 					iconName="plus"
 					label="Add course"
 					onClick={onAddCourseClick}