+++ packages/shared/src/index.ts
@@ -21,6 +21,7 @@ export {
 	getPercentage,
 	getSanitizedHtml,
 	getSizeInBytes,
+	getTimeDistanceFormatDate,
 	initDebounce,
 } from "./libs/helpers/helpers.js";
 export { type Config } from "./libs/modules/config/config.js";
@@ -45,6 +46,19 @@ export {
 	type ValidationSchema,
 	type ValueOf,
 } from "./libs/types/types.js";
+export {
+	ActivitiesApiPath,
+	type ActivityCounts,
+	type ActivityCreateRequestDto,
+	type ActivityDeleteRequestDto,
+	type ActivityGetAllResponseDto,
+	type ActivityPayloadMap,
+	type ActivityResponseDto,
+	ActivityType,
+	activityActionIdParameterValidationSchema,
+	activityCreateFinishSectionValidationSchema,
+	activityDeleteFinishSectionValidationSchema,
+} from "./modules/activities/activities.js";
 export { AuthApiPath, AuthError } from "./modules/auth/auth.js";
 export {
 	type ChatMessageCreateRequestDto,
@@ -64,10 +78,24 @@ export {
 	type ChatGetAllItemResponseDto,
 	type ChatItemResponseDto,
 	type ChatResponseDto,
+	type ChatSearchResponseDto,
 	ChatsApiPath,
 	chatCreateValidationSchema,
 	chatIdParameterValidationSchema,
 } from "./modules/chats/chats.js";
+export {
+	type CommentCreateRequestDto,
+	CommentError,
+	type CommentGetAllRequestDto,
+	type CommentGetAllResponseDto,
+	type CommentUpdateRequestDto,
+	type CommentWithRelationsResponseDto,
+	CommentsApiPath,
+	commentCreateBodyValidationSchema,
+	commentGetAllQueryValidationSchema,
+	commentIdParameterValidationSchema,
+	commentTextValidationSchema,
+} from "./modules/comments/comments.js";
 export {
 	type CourseSectionAddRequestDto,
 	type CourseSectionDto,
@@ -109,6 +137,15 @@ export {
 	friendIdParameterValidationSchema,
 } from "./modules/friends/friends.js";
 export {
+	type AllNotificationsResponseDto,
+	type CreateNotificationRequestDto,
+	NotificationError,
+	type NotificationResponseDto,
+	NotificationStatus,
+	NotificationType,
+	type UpdateNotificationRequestDto,
+} from "./modules/notifications/notifications.js";
+export {
 	SectionStatus,
 	type SectionStatusAddRequestDto,
 	SectionStatusError,
@@ -129,7 +166,13 @@ export {
 	userIdParameterValidationSchema,
 } from "./modules/user-courses/users-courses.js";
 export {
+	type ReadNotificationsRequestDto,
+	UserNotificationsApiPath,
+	readNotificationsRequestValidationSchema,
+} from "./modules/user-notifications/user-notifications.js";
+export {
 	type UserAuthResponseDto,
+	type UserDetailsResponseDto,
 	UserError,
 	type UserGetAllResponseDto,
 	type UserGetByIdRequestDto,