+++ packages/shared/src/modules/notifications/libs/types/all-notifications-response-dto.type.ts
@@ -1,0 +1,7 @@
+import { type NotificationResponseDto } from "./notification-response-dto.type.js";
+
+type AllNotificationsResponseDto = {
+	items: NotificationResponseDto[];
+};
+
+export { type AllNotificationsResponseDto };