+++ packages/shared/src/modules/notifications/notifications.ts
@@ -1,0 +1,8 @@
+export { NotificationStatus, NotificationType } from "./libs/enums/enums.js";
+export { NotificationError } from "./libs/exceptions/exceptions.js";
+export {
+	type AllNotificationsResponseDto,
+	type CreateNotificationRequestDto,
+	type NotificationResponseDto,
+	type UpdateNotificationRequestDto,
+} from "./libs/types/types.js";